Remove CurrencyMintService
