+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/service/state/currency/CurrencyService.java
@@ -12,6 +12,7 @@ import com.apollocurrency.aplwallet.apl.core.entity.state.currency.CurrencySuppl
 import com.apollocurrency.aplwallet.apl.core.entity.state.currency.CurrencyTransfer;
 import com.apollocurrency.aplwallet.apl.core.entity.state.exchange.Exchange;
 import com.apollocurrency.aplwallet.apl.core.transaction.messages.MonetarySystemCurrencyIssuance;
+import com.apollocurrency.aplwallet.apl.core.transaction.messages.MonetarySystemCurrencyMinting;
 
 public interface CurrencyService {
     DbIterator<Currency> getAllCurrencies(int from, int to);
@@ -87,4 +88,8 @@ public interface CurrencyService {
 
     void validateCurrencyNaming(long issuerAccountId, MonetarySystemCurrencyIssuance attachment) throws AplException.ValidationException;
 
+    void mintCurrency(LedgerEvent event, long eventId, Account account,
+                      MonetarySystemCurrencyMinting attachment);
+
+    long getMintCounter(long currencyId, long accountId);
 }
