+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/service/state/currency/impl/MonetaryCurrencyMintingServiceImpl.java
@@ -4,28 +4,21 @@
 
 package com.apollocurrency.aplwallet.apl.core.service.state.currency.impl;
 
-import javax.inject.Inject;
+import com.apollocurrency.aplwallet.apl.core.entity.state.currency.Currency;
+import com.apollocurrency.aplwallet.apl.core.service.state.currency.MonetaryCurrencyMintingService;
+import com.apollocurrency.aplwallet.apl.core.transaction.messages.MonetarySystemCurrencyMinting;
+import com.apollocurrency.aplwallet.apl.crypto.HashFunction;
+
 import javax.inject.Singleton;
 import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.util.Arrays;
 
-import com.apollocurrency.aplwallet.apl.core.entity.state.currency.Currency;
-import com.apollocurrency.aplwallet.apl.core.entity.state.currency.CurrencySupply;
-import com.apollocurrency.aplwallet.apl.core.service.state.currency.CurrencyService;
-import com.apollocurrency.aplwallet.apl.core.service.state.currency.MonetaryCurrencyMintingService;
-import com.apollocurrency.aplwallet.apl.core.transaction.messages.MonetarySystemCurrencyMinting;
-import com.apollocurrency.aplwallet.apl.crypto.HashFunction;
-
 @Singleton
 public class MonetaryCurrencyMintingServiceImpl implements MonetaryCurrencyMintingService {
 
-    private final CurrencyService currencyService;
-
-    @Inject
-    public MonetaryCurrencyMintingServiceImpl(CurrencyService currencyService) {
-        this.currencyService = currencyService;
+    public MonetaryCurrencyMintingServiceImpl() {
     }
 
     @Override
@@ -33,21 +26,11 @@ public class MonetaryCurrencyMintingServiceImpl implements MonetaryCurrencyMinti
         byte[] hash = getHash(currency.getAlgorithm(), attachment.getNonce(), attachment.getCurrencyId(), attachment.getUnits(),
             attachment.getCounter(), accountId);
 
-        Currency currencyFull = this.localSupplyDependency(currency);
-
         byte[] target = getTarget(currency.getMinDifficulty(), currency.getMaxDifficulty(),
             attachment.getUnits(),
-            currencyFull.getCurrencySupply().getCurrentSupply() - currency.getReserveSupply(),
+            currency.getCurrencySupply().getCurrentSupply() - currency.getReserveSupply(),
             currency.getMaxSupply() - currency.getReserveSupply());
         return meetsTarget(hash, target);
-    }
-
-    private Currency localSupplyDependency(Currency currency) {
-        CurrencySupply currencySupply = currencyService.loadCurrencySupplyByCurrency(currency); // load dependency
-        if (currencySupply != null) {
-            currency.setCurrencySupply(currencySupply);
-        }
-        return currency;
     }
 
     @Override
@@ -100,9 +83,8 @@ public class MonetaryCurrencyMintingServiceImpl implements MonetaryCurrencyMinti
 
     @Override
     public BigInteger getNumericTarget(Currency currency, long units) {
-        Currency currencyFull = this.localSupplyDependency(currency);
         return getNumericTarget(currency.getMinDifficulty(), currency.getMaxDifficulty(), units,
-            currencyFull.getCurrencySupply().getCurrentSupply() - currency.getReserveSupply(),
+            currency.getCurrencySupply().getCurrentSupply() - currency.getReserveSupply(),
             currency.getMaxSupply() - currency.getReserveSupply());
     }
